+++ tests/unit/helpers/numbers/basic-math.test.js
@@ -2,7 +2,7 @@
 // BASIC MATH - UNIT TESTS
 // =============================================================================
 
-const numbersHelper = require('../../../../helpers/numbers.helper');
+const numbersHelper = require('../../../../utils/numbers.util');
 
 describe('Basic Math Functions', () => {
   beforeEach(() => {
