+++ tests/unit/helpers/numbers/percentage.test.js
@@ -2,7 +2,7 @@
 // PERCENTAGE - UNIT TESTS
 // =============================================================================
 
-const numbersHelper = require('../../../../helpers/numbers.helper');
+const numbersHelper = require('../../../../utils/numbers.util');
 
 describe('Percentage Functions', () => {
   beforeEach(() => {
