+++ tests/unit/helpers/numbers/random.test.js
@@ -2,7 +2,7 @@
 // RANDOM - UNIT TESTS
 // =============================================================================
 
-const numbersHelper = require('../../../../helpers/numbers.helper');
+const numbersHelper = require('../../../../utils/numbers.util');
 
 describe('Random Functions', () => {
   beforeEach(() => {
