+++ tests/unit/helpers/numbers/utilities.test.js
@@ -2,7 +2,7 @@
 // UTILITIES - UNIT TESTS
 // =============================================================================
 
-const numbersHelper = require('../../../../helpers/numbers.helper');
+const numbersHelper = require('../../../../utils/numbers.util');
 
 describe('Module Exports and Integrations', () => {
   describe('Integration Tests', () => {
