+++ tests/unit/helpers/numbers/validation.test.js
@@ -2,7 +2,7 @@
 // VALIDATION - UNIT TESTS
 // =============================================================================
 
-const numbersHelper = require('../../../../helpers/numbers.helper');
+const numbersHelper = require('../../../../utils/numbers.util');
 
 describe('Validation Functions', () => {
   describe('isValidNumber', () => {
